feat: expand capability of xgb pre fitted model to XGBSEStackedWeibull 1st step
